# ingester: use more efficient spanlogger.DebugLog call