--- a/cmd/mimir/help-all.txt.tmpl
+++ b/cmd/mimir/help-all.txt.tmpl
@@ -2434,7 +2434,7 @@ Usage of ./cmd/mimir/mimir:
   -server.grpc-listen-port int
     	gRPC server listen port. (default 9095)
   -server.grpc-max-concurrent-streams uint
-    	Limit on the number of concurrent streams for gRPC calls (0 = unlimited) (default 100)
+    	Limit on the number of concurrent streams for gRPC calls per client connection (0 = unlimited) (default 100)
   -server.grpc-max-recv-msg-size-bytes int
     	Limit on the size of a gRPC message this server can receive (bytes). (default 104857600)
   -server.grpc-max-send-msg-size-bytes int