#### What this PR does

Reduce overhead of debug logging in ingester push requests, via a change to dskit SpanLogger 
https://github.com/grafana/dskit/pull/367.

I modified the benchmark to send 10 series instead of 10,000 - this is much more like what we get with typical Prometheus senders in a large deployment (500 samples sent x3 replication / 150 ingesters).

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
               │  before.txt  │              after.txt              │
               │    sec/op    │    sec/op     vs base               │
IngesterPush-4   26.24µ ± 21%   21.42µ ± 10%  -18.37% (p=0.002 n=6)

               │  before.txt  │              after.txt              │
               │     B/op     │     B/op      vs base               │
IngesterPush-4   3.021Ki ± 1%   1.119Ki ± 1%  -62.96% (p=0.002 n=6)

               │ before.txt │             after.txt             │
               │ allocs/op  │ allocs/op   vs base               │
IngesterPush-4   51.00 ± 0%   26.00 ± 0%  -49.02% (p=0.002 n=6)
```

#### Checklist

- NA Tests updated
- NA Documentation added
- NA `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`